requirements.txt
python_dateutil==2.8.2
rpaframework==22.5.3
selenium==4.9.1